% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acclogging.r
\name{read_acclogging_csv}
\alias{read_acclogging_csv}
\title{Read a csv file produced by TIMBER or the Java Export CLI}
\usage{
read_acclogging_csv(file, variables = NULL, pattern = NULL)
}
\arguments{
\item{file}{Anything accepted by \code{\link[=read_lines]{read_lines()}} or an xrootd URL.}

\item{variables}{A vector of variables to read (if NULL, read all).}

\item{pattern}{Filter array variables based on regular expression.}
}
\value{
A list of data frames.
}
\description{
Read a csv file produced by TIMBER or the Java Export CLI
}
\examples{
read_acclogging_csv(readcern_example('Summary_4691_B1.HYB.CSV.gz'))
read_acclogging_csv(readcern_example('Summary_4691_B1.HYB.CSV.gz'),
                    variables=c('BSRL.B1.HYB.Q_sat_bunched'))
}
