% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvss.r
\name{read_pvss_export}
\alias{read_pvss_export}
\title{Read a file produced by PVSSExport.}
\usage{
read_pvss_export(file)
}
\arguments{
\item{file}{Anything accepted by \code{\link[=read_csv]{read_csv()}} or an xrootd URL.}
}
\value{
A data frame.
}
\description{
This is a simple wrapper around \code{\link[=read_csv]{read_csv()}}. Columns are renamed
and times are converted to UNIX timestamps.
}
\examples{
read_pvss_export(readcern_example('pvss.csv'))
read_pvss_export('root://eosuser.cern.ch//eos/user/r/rmatev/readcern/pvss.csv')
}
\seealso{
\code{\link[=read_csv]{read_csv()}}.
}
