% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readcern.r
\docType{package}
\name{readcern-package}
\alias{readcern}
\alias{readcern-package}
\title{Read columnar data at CERN.}
\description{
This is a utility package for reading columnar file formats found
at CERN. It provides support for reading over the XRootD protocol.
}
\author{
\strong{Maintainer}: Rosen Matev \email{r.matev@gmail.com}

}
