% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readcern.r
\name{readcern_example}
\alias{readcern_example}
\title{Get path to readcern example}
\usage{
readcern_example(path)
}
\arguments{
\item{path}{Name of file.}
}
\description{
readcern comes bundled with a number of sample files in its inst/extdata
directory. This function make them easy to access.
}
\examples{
readcern_example('pvss.csv')
}
