% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrd.r
\name{xrd}
\alias{xrd}
\alias{xrd_command}
\title{Open an xrootd connection.}
\usage{
xrd(url, decompress = TRUE)

xrd_command(url, decompress = TRUE)
}
\arguments{
\item{url}{URL starting with \code{root://}.}

\item{decompress}{Whether to automatically decompress files.}
}
\value{
A connection or the command to be given to \code{\link[=pipe]{pipe()}}.
}
\description{
\code{xrd()} returns a pipe, while \code{xrd_command()} returns the command
with which an xrootd url is piped.
}
\examples{
xrd('root://eosuser.cern.ch//eos/user/r/rmatev/readcern/README')
xrd_command('root://eosuser.cern.ch//eos/user/r/rmatev/readcern/README')
}
